feat: redirect to team on root page sign in
When you log into the studio at the root endpoint, it doesn't redirect to my profile. This is confusing because as a user, I have to know that I must click the 'Home' button in order to get to the correct route and show Studio functionality. This adds logic to redirect to the personal team route if the user has signed in.