fix: form validation & submission
Forms could no longer be sumbitted (registering username, tables, proj, etc.). If you log the `debouncedValue` in `handleInputChange`, it was *never* updating and always undefined/empty. In this fix, `updateQuery` is called directly when the timeout expires. Also, `debouncedValue` is removed to focus directly on using `inputValue` for the query update (vs. `value`), which is now set via `useEffect`.